+++ packages/blocks/core/js/wordpress/image/test/block.inner-blocks.e2e.cy.js
@@ -6,6 +6,7 @@ import {
 	createPost,
 	appendBlocks,
 	setInnerBlock,
+	setParentBlock,
 	redirectToFrontPage,
 } from '@blockera/dev-cypress/js/helpers';
 
@@ -15,9 +16,9 @@ describe('Image Block → Inner Blocks', () => {
 	});
 
 	it('Inner blocks existence + CSS selectors in editor and front-end', () => {
-		appendBlocks(`<!-- wp:image {"id":7139,"width":"706px","height":"auto","aspectRatio":"1","sizeSlug":"full","linkDestination":"none"} -->
-<figure class="wp-block-image size-full is-resized"><img src="https://placehold.co/600x400" alt="this is the test
-" class="wp-image-7139" style="aspect-ratio:1;width:706px;height:auto"/><figcaption class="wp-element-caption">Image caption is here...</figcaption></figure>
+		appendBlocks(`<!-- wp:image {"id":7139,"width":"706px","height":"auto","aspectRatio":"1","sizeSlug":"full","linkDestination":"custom"} -->
+<figure class="wp-block-image size-full is-resized"><a href="https://placehold.co/600x400"><img src="https://placehold.co/600x400" alt="this is the test
+" class="wp-image-7139" style="aspect-ratio:1;width:706px;height:auto"/></a><figcaption class="wp-element-caption">Image caption is here...</figcaption></figure>
 <!-- /wp:image -->
 		`);
 
@@ -49,6 +50,27 @@ describe('Image Block → Inner Blocks', () => {
 			});
 
 		//
+		// 1.2. Image link inner block
+		//
+		setParentBlock();
+		setInnerBlock('elements/link');
+
+		//
+		// 1.1.1. BG color
+		//
+		cy.setColorControlValue('BG Color', 'ff0000');
+
+		cy.getBlock('core/image')
+			.first()
+			.within(() => {
+				cy.get('a').should(
+					'have.css',
+					'background-color',
+					'rgb(255, 0, 0)'
+				);
+			});
+
+		//
 		// 2. Assert inner blocks selectors in front end
 		//
 		savePage();
@@ -57,6 +79,11 @@ describe('Image Block → Inner Blocks', () => {
 		cy.get('.blockera-block').within(() => {
 			// caption inner block
 			cy.get('figcaption')
+				.first()
+				.should('have.css', 'background-color', 'rgb(255, 0, 0)');
+
+			// link inner block
+			cy.get('a')
 				.first()
 				.should('have.css', 'background-color', 'rgb(255, 0, 0)');
 		});